+++ packages/firebase_messaging/firebase_messaging/ios/Classes/FLTFirebaseMessagingPlugin.m
@@ -948,8 +948,8 @@ NSString *const kMessagingPresentationOptionsUserDefaults =
 
         // Apple only
         // message.notification.apple.badge
-        if (apsAlertDict[@"badge"] != nil) {
-          notificationIOS[@"badge"] = apsAlertDict[@"badge"];
+        if (apsDict[@"badge"] != nil) {
+          notificationIOS[@"badge"] = [NSString stringWithFormat:@"%@", apsDict[@"badge"]];
         }
       }