+++ packages/cloud_firestore_odm/cloud_firestore_odm/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.0-dev.50
+
+ - Update a dependency to the latest release.
+
 ## 1.0.0-dev.49
 
  - Update a dependency to the latest release.