+++ packages/firebase_auth/firebase_auth/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 4.2.9
+
+ - Update a dependency to the latest release.
+
 ## 4.2.8
 
  - Update a dependency to the latest release.