+++ packages/firebase_dynamic_links/firebase_dynamic_links/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 5.0.15
+
+ - Update a dependency to the latest release.
+
 ## 5.0.14
 
  - Update a dependency to the latest release.