+++ packages/firebase_ui_oauth_apple/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.0.22
+
+ - Update a dependency to the latest release.
+
+## 1.0.21
+
+ - Update a dependency to the latest release.
+
 ## 1.0.20
 
  - Update a dependency to the latest release.